--- constructorio-client/src/test/java/io/constructor/client/ConstructorIO_AddBatch_Test.java
@@ -1,36 +1,0 @@
-package io.constructor.client;
-
-import static org.junit.Assert.assertTrue;
-
-import java.util.*;
-
-import org.junit.Test;
-
-public class ConstructorIO_AddBatch_Test {
-
-  public String getRandString() {
-    return UUID.randomUUID().toString().replaceAll("[\\s\\-()]", "");
-  }
-
-  @Test
-  public void addBatchShouldReturn() throws Exception {
-      ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-      ConstructorItem[] items = {
-          new ConstructorItem(this.getRandString(), "Search Suggestions"),
-          new ConstructorItem(this.getRandString(), "Search Suggestions"),
-          new ConstructorItem(this.getRandString(), "Search Suggestions")
-      };
-      assertTrue("batch addition succeeds", constructor.addItemBatch(items, "Search Suggestions"));
-  }
-
-  @Test
-  public void addOrUpdateBatchShouldReturn() throws Exception {
-    ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", true, null);
-    ConstructorItem[] items = {
-        new ConstructorItem(this.getRandString(), "Search Suggestions"),
-        new ConstructorItem(this.getRandString(), "Search Suggestions"),
-        new ConstructorItem(this.getRandString(), "Search Suggestions")
-    };
-    assertTrue("batch upsert succeeds", constructor.addOrUpdateItemBatch(items, "Search Suggestions"));
-  }
-}
