--- constructorio-client/src/test/java/io/constructor/client/ConstructorIO_New_Test.java
@@ -1,44 +1,0 @@
-package io.constructor.client;
-
-import static org.junit.Assert.assertEquals;
-
-import org.junit.Test;
-
-public class ConstructorIO_New_Test {
-
-    @Test
-    public void newShouldSetApiToken() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", true, null);
-        assertEquals("api token should be set", constructor.apiToken, "boinkaToken");
-    }
-
-    @Test
-    public void newShouldSetApiKey() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", true, null);
-        assertEquals("api key should be set", constructor.apiKey, "doinkaKey");
-    }
-
-    @Test
-    public void newShouldSetProtocol() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", false, null);
-        assertEquals("protocol should be set", constructor.protocol, "http");
-    }
-
-    @Test
-    public void newShouldSetHostname() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", false, "com.cnstrc.ac");
-        assertEquals("host should be set", constructor.host, "com.cnstrc.ac");
-    }
-
-    @Test
-    public void newShouldSetHostnameDefault() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("boinkaToken", "doinkaKey", false, null);
-        assertEquals("host should be set to default", constructor.host, "ac.cnstrc.com");
-    }
-
-    @Test
-    public void verifyShouldReturnTrueWithValidKeyTokenPair() throws Exception {
-        ConstructorIO constructor = new ConstructorIO("YSOxV00F0Kk2R0KnPQN8", "ZqXaOfXuBWD4s3XzCI1q", false, null);
-        assertEquals("verify should return true for testing key/pair", constructor.verify(), true);
-    }
-}
